backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
websockets==14.1
python-dotenv==1.0.1
email-validator>=2.0.0

# ML libraries - May need Python 3.10-3.11 for full compatibility
# If scikit-learn fails, install older version or use Python 3.11
scikit-learn>=1.3.0
numpy>=1.24.0,<3.0.0

Pillow==11.0.0
cryptography==44.0.0

cmake>=3.18.0

# Face recognition libraries
# Using dlib-bin (pre-built wheel) instead of compiling from source
# Works with Python 3.11.x - 3.12.x
dlib-bin==20.0.0
# face_recognition==1.3.0
face-recognition-models==0.3.0

# Alternative face detection (lightweight):
# mediapipe==0.10.11
# opencv-python==4.10.0
